+++ src/legacy/Table/stories/TableWithRowSelectAndToolbar.stories.tsx
@@ -1,12 +1,12 @@
 import { Table } from '../Table';
 import { columns } from './columns';
 import { data, type Location } from './data';
-import { Button } from '../..';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus, faCheck, faMinus } from '@fortawesome/free-solid-svg-icons';
+import { Button } from '../../../components';
 
 export default {
-  title: 'Components/Table',
+  title: 'Legacy/Table',
 };
 
 const Example = () => {